Move selected index and search to reports reducer as it belongs to reports page Templates will only need search
