+++ src/modules/templates.js
@@ -5,8 +5,7 @@ import { API_BASE_URL } from '../constants/global';
 const GET_USER_TEMPLATES = 'templates/GET_USER_TEMPLATES';
 const SET_LOADING_TEMPLATES = 'templates/SET_LOADING_TEMPLATES';
 const SET_LOADING_TEMPLATES_ERROR = 'templates/SET_LOADING_TEMPLATES_ERROR';
-const SET_SELECTED_TEMPLATE_INDEX = 'templates/SET_SELECTED_TEMPLATE_INDEX';
-const SET_TEMPLATE_SEARCH_PARAMS = 'templates/SET_TEMPLATE_SEARCH_PARAMSsetTemplateSearchParams';
+const SET_TEMPLATE_SEARCH_PARAMS = 'templates/SET_TEMPLATE_SEARCH_PARAMS';
 
 // Reducer
 const initialState = {
@@ -14,7 +13,6 @@ const initialState = {
   data: {},
   loading: false,
   error: null,
-  selectedIndex: 0,
   searchParams: {}
 };
 
@@ -29,8 +27,6 @@ export default function reducer(state = initialState, action) {
       return Object.assign({}, state, { loading: action.payload });
     case SET_LOADING_TEMPLATES_ERROR:
       return Object.assign({}, state, { error: action.payload });
-    case SET_SELECTED_TEMPLATE_INDEX:
-      return Object.assign({}, state, { selectedIndex: action.payload });
     case SET_TEMPLATE_SEARCH_PARAMS:
       return Object.assign({}, state, { searchParams: action.payload });
     default:
@@ -76,13 +72,6 @@ export function getUserTemplates() {
           payload: false
         });
       });
-  };
-}
-
-export function setSelectedTemplateIndex(index) {
-  return {
-    type: SET_SELECTED_TEMPLATE_INDEX,
-    payload: index
   };
 }
